Update CONTRIBUTING.md
Navigating a new codebase can be challenging. As you're using docs.github.com, you may come across an article that you want to make an update to. You can click on the make a contribution button right on that project, which will take you to the file in this repo where you'll make your changes.